update chat test

+++ apps/jobs/views.py
@@ -9,6 +9,7 @@ from ninja.responses import Response
 from ninja_jwt.authentication import JWTAuth
 
 from jobs import tasks
+from jobs.enums import StageType
 from jobs.models import JobFilter, JobApplication, JobPost, JobApplicationWithdrawal, SavedJob
 from jobs.schemas import TalentJobPostListSchema, TalentJobFilterSchema, \
     TalentJobApplySchema, TalentJobApplicationWithdrawalSchema, ShareJobPostViaEmailSchema
@@ -67,7 +68,7 @@ def update_talent_job_filter(request, data: PatchDict[TalentJobFilterSchema]):
     return Response(status=200, data={"message": "Job filter updated successfully"})
 
 
-@router.post("talent/job-posts/{job_post_id}/apply", response={200: None}, tags=["Talent Jobs"])
+@router.post("talent/job-posts/{job_post_id}/apply", auth=JWTAuth(), response={200: None}, tags=["Talent Jobs"])
 def apply_to_job_post(request, job_post_id:UUID, data: PatchDict[TalentJobApplySchema]):
     user = request.user
     if not hasattr(user, "talent"):
@@ -77,12 +78,12 @@ def apply_to_job_post(request, job_post_id:UUID, data: PatchDict[TalentJobApplyS
         raise HttpError(404, "Job post not found")
     if JobApplication.objects.filter(job_post=job_post, applicant=user.talent).exists():
         raise HttpError(400, "Already applied")
-    JobApplication.objects.create(job_post_id=job_post_id, applicant=user.talent,
+    JobApplication.objects.create(job_post=job_post_id, applicant=user.talent,
                                  **data, match=user.talent.job_match_score(job_post))
     return Response(status=200, data={"message": "Applied successfully"})
 
 
-@router.post("talent/job-posts/applications/{application_id}/withdraw", response={200: None}, tags=["Talent Jobs"])
+@router.post("talent/job-posts/applications/{application_id}/withdraw", auth=JWTAuth(), response={200: None}, tags=["Talent Jobs"])
 def withdraw_job_applications(request, application_id:UUID, data: PatchDict[TalentJobApplicationWithdrawalSchema]):
     user = request.user
     if not hasattr(user, "talent"):
@@ -90,12 +91,14 @@ def withdraw_job_applications(request, application_id:UUID, data: PatchDict[Tale
     application = JobApplication.objects.filter(uid=application_id).first()
     if not application:
         raise HttpError(404, "Application not found")
+    if application.stage:
+        raise HttpError(400, "You cannot withdraw this application at this time")
     JobApplicationWithdrawal.objects.create(job_post=application.job_post,
                                             talent=user.talent, **data)
     application.delete()
     return Response(status=200, data={"message": "Withdrawn successfully"})
 
-@router.post("talent/job-posts/{job_post_id}/share-via-email", response={200: None}, tags=["Talent Jobs"])
+@router.post("talent/job-posts/{job_post_id}/share-via-email", auth=JWTAuth(), response={200: None}, tags=["Talent Jobs"])
 def share_job_post_via_email(request, job_post_id:UUID, data: ShareJobPostViaEmailSchema):
     user = request.user
     if not hasattr(user, "talent"):
@@ -108,7 +111,7 @@ def share_job_post_via_email(request, job_post_id:UUID, data: ShareJobPostViaEma
     ))
     return Response(status=200, data={"message": "Shared successfully"})
 
-@router.post("talent/job-posts/{job_post_id}/save", response={200: None}, tags=["Talent Jobs"])
+@router.post("talent/job-posts/{job_post_id}/save", auth=JWTAuth(), response={200: None}, tags=["Talent Jobs"])
 def save_job(request, job_post_id:UUID):
     user = request.user
     if not hasattr(user, "talent"):
